refactor(tbo): drop tbo_aware param, decide in AR layer
