+++ atom/model_ops/communication_op.py
@@ -6,24 +6,42 @@
 A thin wrapper over the aiter TP all_reduce that centralises the TBO-aware
 routing, so call sites (model AR points such as attention output-proj and MoE
 combine) don't each hand-roll an ``if tbo_aware: tbo_all_reduce else: plain``
-branch. Pass ``tbo_aware=True`` at a site that wants its all_reduce to overlap
-the partner ubatch's compute under Two-Batch Overlap; leave it False (default)
-for the plain reduce.
+branch — nor even thread a ``tbo_aware`` flag through. Whether an all_reduce
+should overlap the partner ubatch's compute is decided here, once, from the
+runtime config via ``_tbo_aware_tp_reduce``.
 
-``tbo_aware`` is a per-site opt-in rather than an internal ``tbo_active()``
-probe on purpose: TBO+DP overlaps via the DP gather/scatter path, not this
-pure-TP reduce, so only sites explicitly on the pure-TP+TBO path should route
-here. The ``tbo_all_reduce`` custom op still no-ops back to a plain reduce when
-TBO is inactive or ``ATOM_TBO_TP_AR_MODE != overlap``.
+Only the pure TP+TBO case routes through the overlap op: TBO+DP overlaps via the
+DP gather/scatter path (not this pure-TP reduce), and non-TBO wants the plain
+reduce with no custom-op / Dynamo-barrier indirection. The ``tbo_all_reduce``
+custom op itself still no-ops back to a plain reduce when TBO is inactive or
+``ATOM_TBO_TP_AR_MODE != overlap``.
 """
 
 import torch
 
 
-def tensor_model_parallel_all_reduce(
-    x: torch.Tensor, tbo_aware: bool = False
-) -> torch.Tensor:
-    """TP all_reduce; routes through the TBO-aware custom op when tbo_aware.
+def _tbo_aware_tp_reduce(tp_size: int) -> bool:
+    """True iff a pure-TP all_reduce should route through the TBO-aware custom op.
+
+    Only the pure TP+TBO case (tp>1, TBO on, no DP) benefits: the ubatch's AR
+    then overlaps the partner ubatch's compute. Non-TBO keeps the plain
+    all_reduce (no custom-op / Dynamo-barrier indirection), and TBO+DP is left
+    untouched (that path overlaps via DP gather/scatter, not this pure-TP AR).
+    """
+    if tp_size <= 1:
+        return False
+    from atom.config import get_current_atom_config
+
+    cfg = get_current_atom_config()
+    return (
+        getattr(cfg, "enable_tbo", False)
+        and cfg.parallel_config.data_parallel_size <= 1
+    )
+
+
+def tensor_model_parallel_all_reduce(x: torch.Tensor) -> torch.Tensor:
+    """TP all_reduce; routes through the TBO-aware custom op on the pure-TP+TBO
+    path (decided internally via ``_tbo_aware_tp_reduce``), else a plain reduce.
 
     The aiter import is lazy (inside the function): this module is pulled in
     very early via atom.model_ops.__init__ -> ... -> linear, before aiter.dist
@@ -31,7 +49,9 @@ def tensor_model_parallel_all_reduce(
     resolves wrong / circularly. Importing at call time (same pattern as
     module_dispatch_ops.tbo_all_reduce) sidesteps that.
     """
-    if tbo_aware:
+    from aiter.dist.parallel_state import get_tp_group
+
+    if _tbo_aware_tp_reduce(get_tp_group().world_size):
         return torch.ops.aiter.tbo_all_reduce(x)
     from aiter.dist.communication_op import tensor_model_parallel_all_reduce as ar
