+++ atom/model_ops/linear.py
@@ -926,9 +926,7 @@ class LinearBase(nn.Module):
                 if self.bias is not None:
                     y += self.bias
         if self.tp_dim == 1 and self.tp_size > 1 and self.reduce_results:
-            y = tensor_model_parallel_all_reduce(
-                y, tbo_aware=getattr(self, "tbo_aware", False)
-            )
+            y = tensor_model_parallel_all_reduce(y)
         return y
 
 
@@ -1822,7 +1820,6 @@ class RowParallelLinear(LinearBase):
         bias: bool = False,
         quant_config: QuantizationConfig | None = None,
         reduce_results: bool = True,
-        tbo_aware: bool = False,
         source_quant_dtype: torch.dtype = None,
         prefix: str = "",
         **kwargs,
@@ -1838,10 +1835,6 @@ class RowParallelLinear(LinearBase):
             source_quant_dtype=source_quant_dtype,
             prefix=prefix,
         )
-        # When True, the built-in TP reduce goes through the TBO-aware custom op
-        # (overlaps with the partner ubatch under TBO). Default off so all other
-        # RowParallelLinear layers keep the plain all_reduce unchanged.
-        self.tbo_aware = tbo_aware
 
     def weight_loader(self, param: nn.Parameter, loaded_weight: torch.Tensor):
         param_data = param.data
